Ensure sequence is moved to CPU
In the TV script generation project, the generate() function erroneously
continues to use the current_seq value without first moving to CPU (if training
on GPU). This can lead to an error of the form shown below:

"TypeError: can't convert CUDA tensor to numpy. Use Tensor.cpu() to copy the
tensor to host memory first."

Resolves: #118